Unify event card with forum thread view
Calendar events used to render in a separate /kalender page with a
stripped-down discussion section that lacked the attachment box, polls,
and other reply-form features. Now event threads use the regular forum
thread page with the event card serving as the first-post slot:

- New EventHeader component takes optional slots (subgroup label, extra
  badges, extra actions, extra menu items) so the event card can embed
  thread context (mute, pin/close) alongside its own actions.
- ThreadPage detects event_slug and renders EventHeader + an info banner
  in place of the implicit OP post; replies and the full reply form
  follow normally.
- /kalender/<slug> redirects community events to the forum thread URL
  (private events without threads keep using the event page).
- create_event_thread no longer seeds a placeholder post; data migration
  0008 drops existing placeholder posts and their FTS index rows.
